Added some checkout shipping address checks that were missing
1. Check if submitting an invalid shipping address to the shipping address step, is failing;
2. Attempt to place an order requiring shipping without providing address (+ fix bug).

### Information about point 2's bug fix:

The checkout summary was checking first the shipping method, then,
the shipping address. Which, verifying the shipping method first was
trying to access a attribute from a shipping address set to None.

The pre-existing test trying to access the checkout summary view
without setting a shipping address was expecting to be redirected to
the shipping method view instead of the shipping address view.

Another pre-existing test was trying to get the checkout summary view
without shipping method but wasn't setting a shipping address.